Create ArduinoServoCommand
This program is my attempt at trying to control a servo using two buttons but one pin. I wanted to do this not only to reduce pin usage which is more efficient. basically I tried storing the rotation of the servo into a variable. This would allow a button to be used twice since the program would command th servo to turn right when the rotation stored is left, and turn left when the rotation stored is right.